requirements.txt
flask
python-dotenv
supabase
flask-bcrypt
validators
gunicorn
flask-session
